refactor(provider): move passport projection write port to application layer
