+++ domain/src/main/java/com/alligator/market/domain/provider/readmodel/passport/projection/projector/ProviderPassportProjector.java
@@ -1,7 +1,7 @@
 package com.alligator.market.domain.provider.readmodel.passport.projection.projector;
 
 import com.alligator.market.domain.provider.passport.ProviderPassport;
-import com.alligator.market.domain.provider.readmodel.passport.projection.port.ProviderPassportProjectionWritePort;
+import com.alligator.market.backend.provider.application.passport.projection.port.out.ProviderPassportProjectionWritePort;
 import com.alligator.market.domain.provider.registry.exception.ProviderRegistryEmptyException;
 import com.alligator.market.domain.provider.vo.ProviderCode;
 import com.alligator.market.domain.provider.registry.ProviderRegistry;
